Experimental implementation of COE redistribution support.

+++ src/x2sys/test_x2sys.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-#	$Id: test_x2sys.sh,v 1.6 2008-10-09 02:05:50 guru Exp $
+#	$Id: test_x2sys.sh,v 1.7 2008-10-10 02:57:09 guru Exp $
 #
 # Test script that exercise the various options in x2sys.
 # We generate a grid and some fake tracks and sample the
@@ -8,7 +8,7 @@
 
 # 1. Make a surface grid with a Mexican hat bump in the middle
 
-delete=1	# Set to 0 for debug where files are not removed
+delete=0	# Set to 0 for debug where files are not removed
 
 grdmath -R-4/4/-4/4 -I0.1 0 0 CDIST DUP DUP MUL NEG 4 DIV EXP EXCH 3 MUL COS MUL = hat.nc
 
@@ -161,7 +161,7 @@ x2sys_list -TFAKE -Cz fake_COE_drift.txt -Fdc -StrackC | psxy -R -J -O -K -Sc0.0
 psxy -R -J -O /dev/null >> $PS
 gv $PS &
 
-# Solve for constants
+# Solve for trends
 x2sys_solve COE.txt -TFAKE -Cz -Ed -V > corr_trend.lis
 
 # Correct tracks
@@ -188,4 +188,7 @@ cut -f3,4 trackCdc.xydz | psxy -R -J -O -K -B5f1/0.2g10WSne -Y-2.25i -W1p,black 
 x2sys_list -TFAKE -Cz fake_COE_drift_corr.txt -Fdc -StrackC | psxy -R -J -O -K -Sc0.05 -Gblack >> $PS
 psxy -R -J -O /dev/null >> $PS
 gv $PS &
-rm -f hat.nc hat.cpt track[ABC]*.xydz fake_COE_*.txt COE.txt corr_const.lis corr_trend.lis
+if [ $delete -eq 1 ]; then
+	rm -f hat.nc hat.cpt track[ABC]*.xydz fake_COE_*.txt COE.txt corr_const.lis corr_trend.lis
+fi
+
